Refactoring: cut MetaflowSettingsTab into several classes for better organization and maintainability.
- Extracted main sections of the MetaFlow settings tab into dedicated classes in `src/settings`:
  - `MetadataInsertionSection`
  - `ExcludeFoldersSection`
  - `FolderFileClassMappingsSection`
  - `PropertyDefaultValueScriptsSection`
  - `SimulationSection`
  - `ExportImportSection`
  - `PluginsStatusSection`
- Created `SettingsUtils` class for shared static utility methods (e.g. checkbox creation, section containers).
- Updated `MetaFlowSettingTab` to delegate rendering and logic to the new section classes, reducing file size and improving readability.
- Migrated drag-and-drop logic, event listeners, and UI controls for folder/fileClass mappings to `FolderFileClassMappingsSection`.
- Refactored note title template management and mapping controls into their dedicated section class.
- Removed legacy display methods from `MetaFlowSettingTab` that are now handled by section classes.
- Ensured all settings functionality (add/edit/remove mappings, property scripts, simulation, import/export, plugin status) is preserved and works as expected.
- Improved maintainability and testability by isolating concerns and reducing code duplication.
- No changes to files in `installScripts` or `bin` folders.
