+++ src/Tests.md
@@ -2,10 +2,10 @@
 
 - [1. Commands](#1-commands)
   - [1.1. Update fileClass](#11-update-fileclass)
-  - [1.2. 1.2 Metadata properties initialization](#12-12-metadata-properties-initialization)
+  - [1.2. Metadata properties initialization](#12-metadata-properties-initialization)
   - [1.3. Sort metadata properties](#13-sort-metadata-properties)
   - [1.4. Move note to the right folder](#14-move-note-to-the-right-folder)
-  - [1.6. Toggle properties panel visibility](#16-toggle-properties-panel-visibility)
+  - [1.5. Toggle properties panel visibility](#15-toggle-properties-panel-visibility)
 - [2. Context menu](#2-context-menu)
   - [2.1. Update metadata in folder](#21-update-metadata-in-folder)
 - [3. Automatic processes](#3-automatic-processes)
@@ -13,10 +13,13 @@
   - [3.2. Create a note from a template with a specific fileClass](#32-create-a-note-from-a-template-with-a-specific-fileclass)
 - [4. Settings](#4-settings)
   - [4.1. General Settings](#41-general-settings)
-  - [4.2. Folder/FileClass Mappings](#42-folderfileclass-mappings)
-  - [4.3. Property Default Value Scripts](#43-property-default-value-scripts)
-  - [4.4. Export/Import Settings](#44-exportimport-settings)
-  - [4.5. Simulation \& Testing](#45-simulation--testing)
+  - [4.2. Plugin dependencies](#42-plugin-dependencies)
+    - [4.2.1. Check no error when templater plugin is disabled](#421-check-no-error-when-templater-plugin-is-disabled)
+    - [4.2.2. Check no error when Metadata Menu plugin is disabled](#422-check-no-error-when-metadata-menu-plugin-is-disabled)
+  - [4.3. Folder/FileClass Mappings](#43-folderfileclass-mappings)
+  - [4.4. Property Default Value Scripts](#44-property-default-value-scripts)
+  - [4.5. Export/Import Settings](#45-exportimport-settings)
+  - [4.6. Simulation \& Testing](#46-simulation--testing)
 
 ## 1. Commands
 
@@ -30,7 +33,7 @@
 - **Edge case:** Run the command on a note with malformed frontmatter. Check for error handling and that the note is not corrupted.
 - **Edge case:** update fileClass from properties panel, should result in the same behavior as updating metadata in the note editor.
 
-### 1.2. 1.2 Metadata properties initialization
+### 1.2. Metadata properties initialization
 
 - disable auto-insert metadata properties in settings.
 - Open a note with frontmatter fileClass.
@@ -58,7 +61,7 @@
 - **Edge case:** Try moving a note when no mapping exists for its fileClass. Confirm no move occurs and a warning is shown.
 - **Edge case:** Move a note that is already in the correct folder. Ensure no duplicate or unnecessary move happens.
 
-### 1.6. Toggle properties panel visibility
+### 1.5. Toggle properties panel visibility
 
 - Open a note in the editor.
 - Use the command or settings to toggle the properties panel.
@@ -108,21 +111,46 @@
 - **Edge case:** Set all toggles off and verify plugin disables all automation.
 - **Edge case:** Enter invalid values in settings fields. Confirm validation and error handling.
 
-### 4.2. Folder/FileClass Mappings
+### 4.2. Plugin dependencies
+
+#### 4.2.1. Check no error when templater plugin is disabled
+
+- open the settings tab.
+- deactivate templater plugin
+- open metaflow settings, check:
+  - if templater is marked as disabled at the bottom of the screen
+  - if import from templater button is disabled and a tooltip is displayed
+- reactivate the plugin
+- reopen metaflow settings
+- the above should work normally after then
+
+#### 4.2.2. Check no error when Metadata Menu plugin is disabled
+
+- open the settings tab.
+- deactivate Metadata Menu plugin
+- open metaflow settings, check:
+  - if Metadata Menu is marked as disabled at the bottom of the screen
+  - if import from Metadata Menu button is disabled and a tooltip is displayed
+- reactivate the plugin
+- reopen metaflow settings
+- the above should work normally after then
+
+
+### 4.3. Folder/FileClass Mappings
 
 - Add, edit, or remove folder/fileClass mappings in settings.
 - Test moving notes and verify they go to the correct folder.
 - **Edge case:** Add duplicate mappings. Confirm only one mapping is used or a warning is shown.
 - **Edge case:** Remove all mappings. Ensure plugin falls back to default behavior.
 
-### 4.3. Property Default Value Scripts
+### 4.4. Property Default Value Scripts
 
 - Add or edit property scripts in settings.
 - Create a note and check if default values are generated by the script.
 - **Edge case:** Add a script with invalid JavaScript. Confirm error is shown and note is not corrupted.
 - **Edge case:** Disable all scripts. Ensure no default values are generated.
 
-### 4.4. Export/Import Settings
+### 4.5. Export/Import Settings
 
 - Use the export button to download settings as JSON.
 - Use the import button to upload a JSON file.
@@ -130,7 +158,7 @@
 - **Edge case:** Import a malformed or incompatible JSON file. Confirm error is shown and settings are not overwritten.
 - **Edge case:** Export settings with empty or default values. Ensure file is valid and can be imported.
 
-### 4.5. Simulation & Testing
+### 4.6. Simulation & Testing
 
 - Use the simulation section in settings to test configuration with sample frontmatter and fileClasses.
 - Verify the simulated output matches expectations.
