+++ src/settings/MetaFlowSettingTab.test.ts
@@ -46,6 +46,12 @@ jest.mock('../externalApi/TemplaterAdapter', () => ({
   }))
 }));
 
+jest.mock('./modals/FileClassAvailableFieldsHelpModal.ts', () => ({
+  FileClassAvailableFieldsHelpModal: jest.fn().mockImplementation(() => ({
+    open: jest.fn()
+  }))
+}));
+
 describe('MetaFlowSettingTab', () => {
   let mockApp: any;
   let mockPlugin: any;
@@ -103,183 +109,7 @@ describe('MetaFlowSettingTab', () => {
     settingTab = new MetaFlowSettingTab(mockApp, mockPlugin);
   });
 
-  describe('Auto-populate from Templater', () => {
-    test('should import folder mappings from Templater settings', async () => {
-      // Clear existing mappings
-      mockPlugin.settings.folderFileClassMappings = [];
-
-      await settingTab['importFolderMappingsFromTemplater']();
-
-      expect(mockPlugin.settings.folderFileClassMappings).toHaveLength(2);
-      expect(mockPlugin.settings.folderFileClassMappings[0]).toEqual({
-        folder: 'Books',
-        fileClass: '',
-        moveToFolder: true,
-        noteTitleTemplates: []
-      });
-      expect(mockPlugin.settings.folderFileClassMappings[1]).toEqual({
-        folder: 'Articles',
-        fileClass: '',
-        moveToFolder: true,
-        noteTitleTemplates: []
-      });
-      expect(mockPlugin.saveSettings).toHaveBeenCalled();
-    });
-
-    test('should not duplicate existing mappings', async () => {
-      // Add existing mapping
-      mockPlugin.settings.folderFileClassMappings = [
-        {folder: 'Books', fileClass: 'existing-book', moveToFolder: false, noteTitleTemplates: []}
-      ];
-
-      await settingTab['importFolderMappingsFromTemplater']();
-
-      // Should still have only one mapping (the existing one)
-      expect(mockPlugin.settings.folderFileClassMappings).toStrictEqual([
-        {"fileClass": "existing-book", "folder": "Books", "moveToFolder": false, "noteTitleTemplates": []},
-        {"fileClass": "", "folder": "Articles", "moveToFolder": true, "noteTitleTemplates": []}
-      ]);
-    });
-
-    test('should handle missing Templater plugin gracefully', async () => {
-      const consoleSpy = jest.spyOn(console, 'error').mockImplementation();
-
-      jest.resetModules();
-      jest.doMock('../externalApi/TemplaterAdapter', () => ({
-        TemplaterAdapter: jest.fn().mockImplementation(async () => ({
-          getFolderTemplatesMapping: jest.fn().mockReturnValue([]),
-        }))
-      }));
-
-      // Re-import after mocking
-      const {MetaFlowSettingTab} = require('./MetaFlowSettingTab');
-      settingTab = new MetaFlowSettingTab(mockApp, mockPlugin);
-
-      await settingTab['importFolderMappingsFromTemplater']();
-
-      // Should not throw error
-      expect(mockPlugin.saveSettings).not.toHaveBeenCalled();
-      consoleSpy.mockRestore();
-    });
-  });
-
-  describe('Auto-populate from MetadataMenu', () => {
-    test('should import property scripts from MetadataMenu settings', async () => {
-      // Setup MetadataMenu plugin mock
-      mockApp.plugins.plugins['metadata-menu'] = {
-        api: {}, // Required for isMetadataMenuAvailable
-        fieldIndex: {
-          fileClassesAncestors: {},
-          fileClassesFields: new Map([
-            ['book', [
-              {name: 'title'},
-              {name: 'author'},
-              {name: 'isbn'}
-            ]],
-            ['article', [
-              {name: 'title'},
-              {name: 'publication'},
-              {name: 'date'}
-            ]]
-          ])
-        },
-        settings: {
-          propertyDefaultValueScripts: [
-            {propertyName: 'title', script: 'return "";', enabled: false, order: 0},
-            {propertyName: 'author', script: 'return "";', enabled: true, order: 1},
-          ]
-        }
-      };
-
-      // Clear existing scripts
-      mockPlugin.settings.propertyDefaultValueScripts = [];
-
-      await settingTab['autoPopulatePropertyScriptsFromMetadataMenu']();
-
-      // Should have imported unique properties
-      expect(mockPlugin.settings.propertyDefaultValueScripts.length).toBeGreaterThan(0);
-
-      // Check that title script exists (used by both fileClasses)
-      const titleScript = mockPlugin.settings.propertyDefaultValueScripts.find(
-        (script: any) => script.propertyName === 'title'
-      );
-      expect(titleScript).toBeDefined();
-      expect(titleScript?.script).toContain('return "";');
-      expect(titleScript?.enabled).toBe(true);
-
-      // Check that author script exists (used by book only)
-      const authorScript = mockPlugin.settings.propertyDefaultValueScripts.find(
-        (script: any) => script.propertyName === 'author'
-      );
-      expect(authorScript).toBeDefined();
-      expect(authorScript?.script).toContain('return "";');
-
-      expect(mockPlugin.saveSettings).toHaveBeenCalled();
-    });
-
-    test('should not duplicate existing property scripts', async () => {
-      // Setup MetadataMenu plugin mock
-      mockApp.plugins.plugins['metadata-menu'] = {
-        settings: {
-          fileClassesFields: {
-            'book': [
-              {name: 'title'}
-            ]
-          }
-        }
-      };
-
-      // Add existing script
-      mockPlugin.settings.propertyDefaultValueScripts = [
-        {propertyName: 'title', script: 'return "existing";', enabled: true, order: 0}
-      ];
-
-      await settingTab['autoPopulatePropertyScriptsFromMetadataMenu']();
-
-      // Should still have only one script (the existing one)
-      expect(mockPlugin.settings.propertyDefaultValueScripts).toHaveLength(2);
-      expect(mockPlugin.settings.propertyDefaultValueScripts[0].script).toBe('return "existing";');
-    });
-
-    test('should handle missing MetadataMenu plugin gracefully', async () => {
-      const consoleSpy = jest.spyOn(console, 'error').mockImplementation();
-      mockPlugin.autoInserter.isMetadataMenuAvailable.mockReturnValue(false);
-      settingTab.metadataMenuAdapter.isMetadataMenuAvailable = jest.fn().mockReturnValue(false);
-
-      await settingTab['autoPopulatePropertyScriptsFromMetadataMenu']();
-
-      // Should not throw error
-      expect(mockPlugin.saveSettings).not.toHaveBeenCalled();
-      consoleSpy.mockRestore();
-    });
-  });
-
   describe('Display Methods', () => {
-    test('should handle empty folder mappings list', () => {
-      const mockContainer = {
-        empty: jest.fn(),
-        createEl: jest.fn().mockReturnValue({
-          createEl: jest.fn().mockReturnValue({
-            style: {},
-            addEventListener: jest.fn(),
-            disabled: false,
-            checked: false,
-            value: ''
-          }),
-          style: {}
-        })
-      };
-
-      mockPlugin.settings.folderFileClassMappings = [];
-
-      // Should not throw error
-      expect(() => {
-        settingTab['displayFolderMappings'](mockContainer as any);
-      }).not.toThrow();
-
-      expect(mockContainer.empty).toHaveBeenCalled();
-    });
-
     test('should toggle autoMoveNoteToRightFolder setting', async () => {
       // Simulate the toggle change
       expect(mockPlugin.settings.autoMoveNoteToRightFolder).toBe(false);
@@ -288,31 +118,6 @@ describe('MetaFlowSettingTab', () => {
       await mockPlugin.saveSettings();
       expect(mockPlugin.settings.autoMoveNoteToRightFolder).toBe(true);
       expect(mockPlugin.saveSettings).toHaveBeenCalled();
-    });
-
-    test('should handle empty property scripts list', () => {
-      const mockContainer = {
-        empty: jest.fn(),
-        createEl: jest.fn().mockReturnValue({
-          createEl: jest.fn().mockReturnValue({
-            style: {},
-            addEventListener: jest.fn(),
-            disabled: false,
-            checked: false,
-            value: ''
-          }),
-          style: {}
-        })
-      };
-
-      mockPlugin.settings.propertyDefaultValueScripts = [];
-
-      // Should not throw error
-      expect(() => {
-        settingTab['displayPropertyScripts'](mockContainer as any);
-      }).not.toThrow();
-
-      expect(mockContainer.empty).toHaveBeenCalled();
     });
 
     describe('Export/Import Settings Section', () => {
